Decouple rendering from generic component
