+++ src/client/components/ToggleSection/PrimaryToggleSection.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import { BLUE, GREY_3, RED, GREEN } from 'govuk-colours'
+import { BLUE, GREY_3, RED } from 'govuk-colours'
 import { FONT_SIZE, FONT_WEIGHTS, SPACING } from '@govuk-react/constants'
 
 import icon from './assets/caret.svg'
@@ -35,7 +35,7 @@ const StyledButton = styled('button')`
     align-items: center;
     background: transparent;
     border: none;
-    font-size: ${({ isStatic }) => (isStatic ? `${FONT_SIZE.SIZE_14}` : `${FONT_SIZE.SIZE_19}`)};
+    font-size: ${FONT_SIZE.SIZE_19};
     color: ${BLUE};
     cursor: pointer;
     font-weight: ${FONT_WEIGHTS.regular};
@@ -66,34 +66,12 @@ const BadgeContainer = styled('span')`
   margin-left: ${SPACING.SCALE_1};
 `
 
-const StyledLabel = styled('label')`
-  display: inline-table;
-  background: transparent;
-  border: none;
-  font-size: 19px;
-  float: right;
-  margin: 5px;
-  color: ${(props) => props.color};
-`
-
-const renderRequiredFields = (requiredFieldFlag, fieldCount) => {
-  if (requiredFieldFlag) {
-    if (fieldCount == 0) {
-      return <StyledLabel color={GREEN}>Completed</StyledLabel>
-    }
-    return <StyledLabel color={RED}>{fieldCount} fields required</StyledLabel>
-  }
-}
-
 const PrimaryToggleSection = ({
   label,
   badge = null,
   open,
   isOpen = false,
   children,
-  showRequiredField = false,
-  fieldCount = 0,
-  isStatic = false,
   ...props
 }) => (
   <ToggleContainer {...props}>
@@ -107,7 +85,6 @@ const PrimaryToggleSection = ({
           {label}
         </ButtonContent>
       </StyledButton>
-      {renderRequiredFields(showRequiredField, fieldCount)}
       {badge && <BadgeContainer>{badge}</BadgeContainer>}
     </StyledHeader>
     <StyledContent isOpen={isOpen}>{children}</StyledContent>
